[REFACTOR] Refactor validators to validators.py, keep consistency in urls.py.
